Pull CoseDetachedSignatures into a dedicated project with nuget dependencies on CoseSign1 nuget package to reduce confusion
