+++ CoseDetachedSIgnature.Tests/Usings.cs
@@ -1,0 +1,16 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+global using System.Security.Cryptography;
+global using System.Security.Cryptography.Cose;
+global using System.Security.Cryptography.X509Certificates;
+global using CoseDetachedSignature;
+global using CoseDetachedSignature.Extensions;
+global using CoseSign1;
+global using CoseSign1.Abstractions.Interfaces;
+global using CoseSign1.Interfaces;
+global using CoseSign1.Tests.Common;
+global using FluentAssertions;
+global using Moq;
+global using NUnit.Framework;
+
