+++ CoseDetachedSignature/Usings.cs
@@ -1,0 +1,19 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+global using System;
+global using System.Collections.Concurrent;
+global using System.Collections.Generic;
+global using System.Diagnostics;
+global using System.IO;
+global using System.Linq;
+global using System.Reflection;
+global using System.Security.Cryptography;
+global using System.Security.Cryptography.Cose;
+global using System.Text.RegularExpressions;
+global using System.Threading.Tasks;
+global using CoseDetachedSignature.Extensions;
+global using CoseSign1.Abstractions.Interfaces;
+global using CoseSign1.Extensions;
+global using CoseSign1.Interfaces;
+global using CoseSign1;
