+++ Nuspec/CoseSign1.Abstractions.nuspec
