+++ Nuspec/CoseSign1.Certificates.nuspec
