+++ Nuspec/CoseSign1.nuspec
